Restructure test fixtures a bit.
