+++ src/eligibility_signposting_api/views/eligibility.py
@@ -10,12 +10,12 @@ from wireup import Injected
 
 from eligibility_signposting_api.model.eligibility import EligibilityStatus, NHSNumber, Status
 from eligibility_signposting_api.services import EligibilityService, UnknownPersonError
-from eligibility_signposting_api.views import response_models
+from eligibility_signposting_api.views.response_model import eligibility
 
 STATUS_MAPPING = {
-    Status.actionable: response_models.Status.actionable,
-    Status.not_actionable: response_models.Status.not_actionable,
-    Status.not_eligible: response_models.Status.not_eligible,
+    Status.actionable: eligibility.Status.actionable,
+    Status.not_actionable: eligibility.Status.not_actionable,
+    Status.not_eligible: eligibility.Status.not_eligible,
 }
 
 logger = logging.getLogger(__name__)
@@ -48,23 +48,23 @@ def check_eligibility(nhs_number: NHSNumber, eligibility_service: Injected[Eligi
 
 def build_eligibility_response(
     eligibility_status: EligibilityStatus,
-) -> response_models.EligibilityResponse:
+) -> eligibility.EligibilityResponse:
     """Return an object representing the API response we are going to send, given an evaluation of the person's
     eligibility."""
-    return response_models.EligibilityResponse(  # pyright: ignore[reportCallIssue]
+    return eligibility.EligibilityResponse(  # pyright: ignore[reportCallIssue]
         response_id=uuid.uuid4(),  # pyright: ignore[reportCallIssue]
-        meta=response_models.Meta(last_updated=response_models.LastUpdated(datetime.now(tz=UTC))),  # pyright: ignore[reportCallIssue]
+        meta=eligibility.Meta(last_updated=eligibility.LastUpdated(datetime.now(tz=UTC))),  # pyright: ignore[reportCallIssue]
         processed_suggestions=[  # pyright: ignore[reportCallIssue]
-            response_models.ProcessedSuggestion(  # pyright: ignore[reportCallIssue]
-                condition_name=response_models.ConditionName(condition.condition_name),  # pyright: ignore[reportCallIssue]
+            eligibility.ProcessedSuggestion(  # pyright: ignore[reportCallIssue]
+                condition_name=eligibility.ConditionName(condition.condition_name),  # pyright: ignore[reportCallIssue]
                 status=STATUS_MAPPING[condition.status],
-                status_text=response_models.StatusText(f"{condition.status}"),  # pyright: ignore[reportCallIssue]
+                status_text=eligibility.StatusText(f"{condition.status}"),  # pyright: ignore[reportCallIssue]
                 eligibility_cohorts=[],  # pyright: ignore[reportCallIssue]
                 suitability_rules=[  # pyright: ignore[reportCallIssue]
-                    response_models.SuitabilityRule(  # pyright: ignore[reportCallIssue]
-                        type=response_models.RuleType(reason.rule_type.value),  # pyright: ignore[reportCallIssue]
-                        rule_code=response_models.RuleCode(reason.rule_name),  # pyright: ignore[reportCallIssue]
-                        rule_text=response_models.RuleText(reason.rule_result),  # pyright: ignore[reportCallIssue]
+                    eligibility.SuitabilityRule(  # pyright: ignore[reportCallIssue]
+                        type=eligibility.RuleType(reason.rule_type.value),  # pyright: ignore[reportCallIssue]
+                        rule_code=eligibility.RuleCode(reason.rule_name),  # pyright: ignore[reportCallIssue]
+                        rule_text=eligibility.RuleText(reason.rule_result),  # pyright: ignore[reportCallIssue]
                     )
                     for reason in condition.reasons
                 ],  # pyright: ignore[reportCallIssue]
